Bugfix: invalid file descptor error on stream close.
This caused both flushing to fail and closing to fail.

Make flushing more flexible and more optional, expanding the close functions to be non-autoflushing and autoflushing.

The fclose() closes the file stream and therefore the descriptor will be invalid.
In this case, the descriptor need only be reset to -1.

If the stream property is not populated and "complete" is designated, then close the file by the file descriptor id.