Update: generalize the pid file creation and deletion.
The PID create and delete functions were written as intended specifically for the main controller program.

PID files are to be used by services and utilities, therefore the PID create and delete functions should also be used there.

Detect if the main PID file is created and then only attempt to delete it on exit (and then do not do so for child processes).