Cleanup: Relocate headers into appropriate sub-directories.
Have the FLL project contain its headers within a sub-directory of 'fll'.
This promotes better organization.

The level_3 programs are just that, programs, and as such will not be under 'fll' but instead will be under their own sub-directory.

The programs are moved out of level_3 and into the program include directory, with their sub-directory using the program name.