Update: FSS Basic Read --delimit parameter improvements.
The previous implementation is weak in that there is no good way to just delimit Object or just delimit Content.

Redesign to allow for specifying the delimit parameter multiple times and therefore allow for customizing what to specify.
Rename "depth" to "content" in the delimit enum to better communicate that this is for "content" delimiting.

Examples:
  - "fss_basic_read --delimit object": Results in delimited Objects but not Content.
  - "fss_basic_read --delimit 0+": Results in delimited Content (position 0 and greater) but not Objects.
  - "fss_basic_read --delimit object --delimit 1-": Results in delimited Objects and delimit Content (position 1 or less).

For this standard, there is no delimit support in Content so the use of the numeric range is superfluous.
Having this functionality, however, makes it consistent with the rest of the FSS Read programs.