Regression: "fake skeleton" no longer works due to "file not found".
ERROR: Unable to create file 'data/build/defines', could not find file.

This error message is a bit nonsensical because it is supposed to create a file if it does not exist.

This is a regression in that at some point f_file_is() started returning F_file_found_not with error bit set and the skeleton program is only expecting F_file_found_not without the error bit set.
Make the test error-bit neutral.